third_party/blink/web_tests/platform/linux-chrome/external/wpt/reporting/document-reporting-not-batch-different-document.https-expected.txt

This is a testharness.js-based test.
[FAIL] Reports are not batched for same url in different document.
  assert_equals: Number of reports expected 2 but got 0
Harness: the test ran to completion.
